refactor: replace design token StoryBoard with ColorToken story
